Simplify structure
