+++ front/src/dashboard/transport/RouteSlipsToCollect.tsx
@@ -3,14 +3,15 @@ import { useParams } from "react-router-dom";
 import { useQuery, gql } from "@apollo/client";
 import { Query } from "generated/graphql/types";
 import { formSearchResultFragment } from "common/fragments";
-import { FormSearchResultTable } from "../../FormSearchResultTable";
+import { Breadcrumb, BreadcrumbItem } from "common/components";
+import { FormSearchResultTable } from "../FormSearchResultTable";
 
-export const SEARCH_COLLECTED = gql`
+export const SEARCH_TO_COLLECT = gql`
   query SearchFollows($siret: String!) {
     searchForms(
       siret: $siret
       status: ["SEALED", "SENT", "RESEALED", "RESENT"]
-      waitingForMe: false
+      waitingForMe: true
     ) {
       ...FormSearchResultFragment
     }
@@ -19,9 +20,9 @@ export const SEARCH_COLLECTED = gql`
 `;
 
 // TODO: cards display (persisted in local storage)
-export default function CollectedTab() {
+export function RouteSlipsToCollect() {
   const { siret } = useParams<{ siret: string }>();
-  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_COLLECTED, {
+  const { data } = useQuery<Pick<Query, "searchForms">>(SEARCH_TO_COLLECT, {
     variables: {
       siret,
     },
@@ -30,5 +31,15 @@ export default function CollectedTab() {
   // TODO: blankslate
   // TODO: loading state
 
-  return <FormSearchResultTable searchResults={data?.searchForms ?? []} />;
+  return (
+    <>
+      <Breadcrumb>
+        <BreadcrumbItem>Transport</BreadcrumbItem>
+        <BreadcrumbItem>
+          Chargés, en attente de réception ou de transfert
+        </BreadcrumbItem>
+      </Breadcrumb>
+      <FormSearchResultTable searchResults={data?.searchForms ?? []} />
+    </>
+  );
 }
