+++ front/src/dashboard/transport/index.ts
@@ -1,0 +1,2 @@
+export * from "./RouteSlipsCollected";
+export * from "./RouteSlipsToCollect";
